When I had a device halted under pyOCD, the CPU usage on my development
machine would sky rocket.  The effect of this was very noticeable when
running off of batteries.  This was due to pyOCD's usage of
non-blocking socket reads in a tight loop while it tries to read
packets sent from GDB.  I added a sleep() call to the run() method as
already used in the resume() method.  I set both of these sleeps to be
100 milliseconds. For me, I didn't find this decreased the perceived
responsiveness of pyOCD when I was using GDB and it greatly reduces the
CPU usage.

**Edit**: Below feature is no longer required after PR #93 so it has been
removed from this PR.

~~As I have been working to implement pyOCD support for the LPC4330, I
would hit issues where resume() would hit an excpeption as it tried to
read the hard fault handler address from the interrupt vector table at
an address of 0x12.  This happened because the external SPIFI FLASH
isn't always available and mapped into this shadow region.  I added
code which handles this exception and treats it the same as when no
hardware breakpoints are available.~~