@c1728p9 2015-01-10
I use eclipse (KDS) as a front end to GDB and this patch makes stepping noticeable slower for me.  Below is some timing information I recorded on my machine.  I used the same test C code for both runs, and recorded the time it took to step over the first two functions.

Before Patch:
Step - 33 gdb commands, 0.809000015 s to complete
Step - 32 gdb commands, 0.80099988 s to complete

With patch:
Step - 33 gdb commands, 4.209000111 s to complete
Step - 32 gdb commands, 4.003999949 s to complete


Would it be possible to modify your patch to only start sleeping a fixed amount of time after the last command completed?  I think this would keep CPU usage at a minimum, and also keep the faster performance when using eclipse.  Something like the below code:

```
while True:
    ...
    if time() - last_command > 0.5
        # Sleep if no command has been sent for 500ms
        sleep(0.1)
    ... 
    if command:
        ...
        last_command = time()
    ...
```