@adamgreen 2015-01-10
@c1728p9 Thanks for the feedback.  I will incorporate your sleep change later today.

I would be happy to see you cleanup that hard fault code.  I had switching to the vector catch on my to do list but if you do it then that is even better :)