@adamgreen 2015-01-10
@c1728p9 I have added the back-offs that you mentioned in your above comments.  Hopefully it helps with your single stepping issue with this commit.