@c1728p9 2015-01-11
I'll give this patch a try and let you know.  Also, I have the vector catch implemented in #93.