@c1728p9 2015-01-11
Looks like this fixes the eclipse slow downs.  I looks like its even a little faster.  Thanks for fixing this.

Before Patch:
Step - 31 gdb commands, 0.811000109s to complete
Step - 30 gdb commands, 0.789000034s to complete

With patch:
Step - 31 gdb commands, 0.784999847s to complete
Step - 30 gdb commands, 0.778999805s to complete