@adamgreen 2015-01-21
Rebased and retested on NRF51822 and LPC1768.